# RRTS/Endpoints/api-docs.json
{"openapi":"3.0.1","info":{"title":"Road Repair and Tracking System API","description":"API documentation for the Road Repair and Tracking System","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Complaints","description":"Endpoints related to complaints"},{"name":"Materials","description":"Endpoints related to materials"},{"name":"User","description":"Endpoints related to users"},{"name":"Resources","description":"Endpoints related to resources"},{"name":"Schedules","description":"Endpoints related to schedules"},{"name":"Statistics","description":"Endpoints related to mayor services"},{"name":"Estimates","description":"Endpoints related to estimates"},{"name":"Authentication","description":"Endpoints related to authentication"}],"paths":{"/api/areas/{id}":{"get":{"tags":["area-controller"],"operationId":"getAreaById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Area"}}}}}},"put":{"tags":["area-controller"],"operationId":"updateArea","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Area"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Area"}}}}}},"delete":{"tags":["area-controller"],"operationId":"deleteArea","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/auth/login":{"post":{"tags":["Authentication"],"summary":"Login","description":"Username/Password based login  system","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/resources":{"get":{"tags":["Resources"],"summary":"Get Resources","description":"Retrieves a list of resources","operationId":"getAllResources","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Resource"}}}}}}},"post":{"tags":["Resources"],"summary":"Create Resource","description":"Creates a resource and adds to DB","operationId":"createResource","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Resource"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Resource"}}}}}}},"/api/materials":{"get":{"tags":["Materials"],"summary":"All Materials","description":"Retireves a list of all materials","operationId":"getAllMaterials","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Material"}}}}}}},"post":{"tags":["Materials"],"summary":"Create Material","description":"Adds a new material to DB","operationId":"createMaterial","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Material"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Material"}}}}}}},"/api/estimates/batch":{"post":{"tags":["Estimates"],"summary":"Batch Submit","description":"Submits all estimates for a complaint","operationId":"createBatchEstimates","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstimateRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/complaints/add":{"post":{"tags":["Complaints"],"summary":"Add complaints","description":"Adds new complaints to the database","operationId":"createComplaint","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/areas":{"get":{"tags":["area-controller"],"operationId":"getAllAreas","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Area"}}}}}}},"post":{"tags":["area-controller"],"operationId":"createArea","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Area"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Area"}}}}}}},"/api/access-logs":{"get":{"tags":["access-log-controller"],"operationId":"getAllAccessLogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccessLog"}}}}}}},"post":{"tags":["access-log-controller"],"operationId":"createAccessLog","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessLog"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccessLog"}}}}}}},"/api/schedules/update/{id}":{"patch":{"tags":["Schedules"],"summary":"Update schedule","description":"Updates a schedule based on UUID","operationId":"updateScheduleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ScheduleRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/resources/update/{id}":{"patch":{"tags":["Resources"],"summary":"Update Resource","description":"Updates a resource based on UUID","operationId":"updateResource","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResourceDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/materials/update/{id}":{"patch":{"tags":["Materials"],"summary":"Update material","description":"Updates a material based on UUID","operationId":"updateResource_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MaterialDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/complaints/update/{id}":{"patch":{"tags":["Complaints"],"summary":"Update complaints","description":"Updates a complaint given the complaint id","operationId":"updateComplaint","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ComplaintReviewDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseMessage"}}}}}}},"/api/users":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Users"}}}}}}}},"/api/users/{id}":{"get":{"tags":["User"],"summary":"Get a single user details","description":"Retrieve a single user","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Users"}}}}}}},"/api/statistics/utilization":{"get":{"tags":["Statistics"],"summary":"Resource Utilization","description":"Retrieves resource utilization of personnel and machinery","operationId":"getResourceUtilization","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResourceUtilizationDTO"}}}}}}},"/api/statistics/repairs":{"get":{"tags":["Statistics"],"summary":"Repairs By Period","description":"Retrieves number and types of completed repairs in given duration","operationId":"getRepairsByPeriod","parameters":[{"name":"startDate","in":"query","required":true,"schema":{"type":"string","format":"date"}},{"name":"endDate","in":"query","required":true,"schema":{"type":"string","format":"date"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RepairStatisticsDTO"}}}}}}}},"/api/statistics/outstanding-repairs":{"get":{"tags":["Statistics"],"summary":"Outstanding Repairs","description":"Retrieves list of pending complaints or repair works","operationId":"getOutstandingRepairs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Schedule"}}}}}}}},"/api/schedules/{id}":{"get":{"tags":["Schedules"],"summary":"Get Schedules","description":"Fetches all schedules","operationId":"getScheduleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Schedule"}}}}}}},"/api/schedules/complaint/{id}":{"get":{"tags":["Schedules"],"summary":"Get Schedule","description":"Gets a schedule for a complaint","operationId":"getScheduleByComplaintId","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/resources/{id}":{"get":{"tags":["Resources"],"summary":"Get Resource","description":"Fetches a single resource based on UUID","operationId":"getResourceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Resource"}}}}}}},"/api/materials/{id}":{"get":{"tags":["Materials"],"summary":"Single Material","description":"Retrieves single material details","operationId":"getMaterialById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Material"}}}}}}},"/api/estimates":{"get":{"tags":["Estimates"],"summary":"Get all estimates","description":"Retrieve a list of all estimates","operationId":"getAllEstimates","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Estimate"}}}}}}}},"/api/estimates/{id}":{"get":{"tags":["Estimates"],"summary":"Single estimate","description":"Retrieve a single estimate","operationId":"getEstimateById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Estimate"}}}}}},"delete":{"tags":["Estimates"],"summary":"Delete Estimate","description":"Deletes an estimate based on UUID","operationId":"deleteEstimate","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/complaints/view/{id}":{"get":{"tags":["Complaints"],"summary":"Get complaint","description":"Retrieve a single complaint","operationId":"getComplaintById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Complaint"}}}}}}},"/api/complaints/complaintList":{"get":{"tags":["Complaints"],"summary":"Get all complaints","description":"Retrieve a list of all complaints","operationId":"getAllComplaints","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Complaint"}}}}}}}},"/api/complaints/complaintList/{areaId}":{"get":{"tags":["Complaints"],"summary":"Area-wise complaints","description":"Retrieve a list of all complaints area-wise","operationId":"getComplaintsByArea","parameters":[{"name":"areaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Complaint"}}}}}}}},"/api/access-logs/{id}":{"get":{"tags":["access-log-controller"],"operationId":"getAccessLogById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccessLog"}}}}}},"delete":{"tags":["access-log-controller"],"operationId":"deleteAccessLog","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK"}}}},"/api/resources/delete/{id}":{"delete":{"tags":["Resources"],"summary":"Delete Resource","description":"Deletes a resource based on UUID","operationId":"deleteResource","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/materials/delete/{id}":{"delete":{"tags":["Materials"],"summary":"Delete Material","description":"Deletes an material based on UUID","operationId":"deleteMaterial","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"Area":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"areaName":{"type":"string"}}},"AuthRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Resource":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"name":{"type":"string"},"count":{"type":"integer","format":"int32"},"type":{"type":"string"},"cost":{"type":"number","format":"double"},"available":{"type":"integer","format":"int32"}}},"Material":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"type":{"type":"string"},"description":{"type":"string"},"inventory":{"type":"integer","format":"int32"},"cost":{"type":"number","format":"double"}}},"EstimateRequestDTO":{"type":"object","properties":{"complaintId":{"type":"string","format":"uuid"},"resources":{"type":"array","items":{"$ref":"#/components/schemas/ResourceEstimateDTO"}},"materials":{"type":"array","items":{"$ref":"#/components/schemas/MaterialEstimateDTO"}}}},"MaterialEstimateDTO":{"type":"object","properties":{"materialId":{"type":"string","format":"uuid"},"quantity":{"type":"integer","format":"int32"}}},"ResourceEstimateDTO":{"type":"object","properties":{"resourceId":{"type":"string","format":"uuid"},"quantity":{"type":"integer","format":"int32"}}},"ResponseMessage":{"type":"object","properties":{"status":{"type":"string"},"message":{"type":"string"}}},"ComplaintDTO":{"type":"object","properties":{"description":{"type":"string"},"roadId":{"type":"integer","format":"int64"},"complaintDate":{"type":"string"},"repairType":{"type":"string"},"residentName":{"type":"string"},"residentContact":{"type":"string"}}},"AccessLog":{"type":"object","properties":{"logId":{"type":"integer","format":"int64"},"user":{"$ref":"#/components/schemas/Users"},"loginTime":{"type":"string","format":"date-time"},"logoutTime":{"type":"string","format":"date-time"}}},"Users":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"fullName":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"branch":{"type":"string"},"area":{"$ref":"#/components/schemas/Area"}}},"ScheduleRequestDTO":{"type":"object","properties":{"startDate":{"type":"string"},"endDate":{"type":"string"},"status":{"type":"string"}}},"ResourceDTO":{"type":"object","properties":{"count":{"type":"integer","format":"int32"},"available":{"type":"integer","format":"int32"},"cost":{"type":"number","format":"double"}}},"MaterialDTO":{"type":"object","properties":{"inventory":{"type":"integer","format":"int32"},"cost":{"type":"number","format":"double"}}},"ComplaintReviewDTO":{"type":"object","properties":{"status":{"type":"string"},"severity":{"type":"integer","format":"int32"},"priority":{"type":"integer","format":"int32"},"supervisorId":{"type":"string","format":"uuid"}}},"ResourceUtilizationDTO":{"type":"object","properties":{"personnelUtilizationPercentage":{"type":"number","format":"double"},"machineryUtilizationPercentage":{"type":"number","format":"double"}}},"RepairStatisticsDTO":{"type":"object","properties":{"repairType":{"type":"string"},"count":{"type":"integer","format":"int64"}}},"Complaint":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"complaintDate":{"type":"string","format":"date"},"road":{"$ref":"#/components/schemas/Road"},"repairType":{"type":"string"},"description":{"type":"string"},"residentName":{"type":"string"},"residentContact":{"type":"string"},"status":{"type":"string"},"severity":{"type":"integer","format":"int32"},"priority":{"type":"integer","format":"int32"},"supervisor":{"$ref":"#/components/schemas/Users"}}},"Road":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roadName":{"type":"string"},"localityType":{"type":"string"},"area":{"$ref":"#/components/schemas/Area"},"additionalInfo":{"type":"string"}}},"Schedule":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"complaint":{"$ref":"#/components/schemas/Complaint"},"scheduledDate":{"type":"string","format":"date"},"startDate":{"type":"string","format":"date"},"deadline":{"type":"string","format":"date"},"endDate":{"type":"string","format":"date"},"status":{"type":"string"}}},"Estimate":{"type":"object","properties":{"estimateId":{"type":"string","format":"uuid"},"complaint":{"$ref":"#/components/schemas/Complaint"},"resource":{"$ref":"#/components/schemas/Resource"},"material":{"$ref":"#/components/schemas/Material"},"quantity":{"type":"integer","format":"int32"}}}}}}